refactor: 2단계 피드백 적용
- 경로 깨진 이미지 파일 확인
- 스크립트 한개 내에서 유저 플로우를 정의하여 여러 화면/기능에 대한 check
- Load Test 의 테스트 시간을 30분 ~ 2시간 사이로 테스트(1시간)
- Load, Stress 테스트의 threshold 시간을 낮춰볼 것
